fix: handing errors

+++ apps/cloud/src/api/permissions.rs
@@ -128,8 +128,8 @@ pub async fn invite_member(
         };
 
         let Ok(addr) = data.email.clone().parse() else {
-        return ErrorStatus::BadRequest.into_response()
-    };
+            return ErrorStatus::BadRequest.into_response()
+        };
 
         let (permission_id, user_cred) = match ctx
             .db
@@ -341,12 +341,12 @@ pub async fn remove_user(
         }
     };
 
-    let permission_model = ctx
-        .db
-        .get_permission_by_id(id.clone())
-        .await
-        .unwrap()
-        .unwrap();
+    let permission_result = ctx.db.get_permission_by_id(id.clone()).await;
+    if permission_result.is_err() || permission_result.as_ref().unwrap().is_none() {
+        return ErrorStatus::InternalServerError.into_response();
+    }
+    let permission_model = permission_result.unwrap().unwrap();
+
     match ctx.db.delete_permission(id).await {
         Ok(true) => {
             if let Some(user_id) = permission_model.user_id {
